refactor: api
userRoutes and config inside the api directory
